separate the sql implementation in different file
