--- src/ai/prompts/ob1/4.ts
@@ -1,3 +1,0 @@
-export const PROMPT = `
-
-`
