--- src/ai/prompts/text-to-worker.ts
@@ -1,3 +1,0 @@
-export const PROMPT = `
-    Return a Cloudflare worker function code snippet based on the user prompt. Only ever return a single code snippet. Do not respond with any additional text or information.
-`
