+++ src/index.ts
@@ -1,7 +1,5 @@
 export * from './connections';
-// export * from './connections/outerbase.backup';
 export * from './connections/sqlite/cloudflare';
-// export * from './connections/neon-http.bk';
 export * from './connections/motherduck';
 export * from './connections/bigquery';
 export * from './connections/sqlite/starbase';
@@ -10,5 +8,4 @@ export * from './connections/mysql';
 export * from './connections/postgre/postgresql';
 export * from './connections/sqlite/turso';
 export * from './client';
-// export * from './models';
 export * from './models/decorators';
