+++ tsconfig.json
@@ -12,11 +12,7 @@
         "experimentalDecorators": true,
         "emitDecoratorMetadata": true,
         "strict": true,
-        "baseUrl": "./",
-        "paths": {
-            "src/*": ["src/*"],
-            "src": ["src"]
-        }
+        "baseUrl": "./"
     },
     "include": ["src/**/*"],
     "exclude": ["node_modules"]
